pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tool-lab"
version = "0.1.0"
description = "Config-driven Tool-Lab experiment runner for fixed and scrolling Mouselab-style tasks."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.51.0",
  "openai>=1.75.0",
  "PyYAML>=6.0.2",
  "google-genai",
  "pydantic>=2.8.0",
  "pandas",
]

[project.scripts]
tool-lab = "tool_lab.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
